fix: API Gateway integration method
Integration method `ANY` is not supported. The reason you can't see it on the AWS Console is because it always defaults to `POST`.
The has nothing to do with the HTTP method used to call the API and is just how API Gateway calls the Lambda API (the handler will still receive the correct method).